Update file moving script to use backup folder
Add 2 files for it to work, .source-folder and .dest-folder.
As its content I added the source and destination folder for all the
files.

dest-folder is a Google Drive folder, which will be backup automatically
when new files are moved in.

At the end, all files in the dest-folder are symbolic linked to data
folder where jupyter notebooks will read them.